+++ components/abstractions/src/main/java/com/microsoft/kiota/ResponseHandler.java
@@ -17,6 +17,7 @@ public interface ResponseHandler {
      * @param errorMappings        the error mappings for the response to use when deserializing failed responses bodies. Where an error code like 401 applies specifically to that status code, a class code like 4XX applies to all status codes within the range if an the specific error code is not present.
      * @param <NativeResponseType> The type of the native response object.
      * @param <ModelType>          The type of the response model object.
+     * @return The deserialized response model object.
      */
     @Nullable
     <NativeResponseType, ModelType> ModelType handleResponse(@Nonnull final NativeResponseType response, @Nullable final HashMap<String, ParsableFactory<? extends Parsable>> errorMappings);